+++ levels/house_ceremony/conversations/corridor-lovelace.pug
@@ -1,12 +1,12 @@
 -
-  const STATE_KEY = 'TQ_API_ACADEMY_WORLD_STATE';
+  const STATE_KEY = 'TQ_HOUSE_CEREMONY_WORLD_STATE';
   const worldState = getState(STATE_KEY);    
 
 mixin excavation
   response#excavation.
     Why does the Academy have an excavation crew?
 
-conversation(display_name="Door to Secret Hopper Storage Room")
+conversation(display_name="Lovelace ")
   case lastResponse
     
     when "excavation"
