+++ levels/house_ceremony/conversations/corridor-neumann.pug
@@ -1,5 +1,5 @@
 -
-  const STATE_KEY = 'TQ_API_ACADEMY_WORLD_STATE';
+  const STATE_KEY = 'TQ_HOUSE_CEREMONY_WORLD_STATE';
   const worldState = getState(STATE_KEY);    
 
 mixin excavation
