+++ levels/house_ceremony/conversations/head-taifa.pug
@@ -1,0 +1,87 @@
+-
+  const STATE_KEY = 'TQ_HOUSE_CEREMONY_WORLD_STATE';
+  const worldState = getState(STATE_KEY);    
+
+mixin unique
+  response#unique.
+    Could you remind me what makes House Turing unique?
+
+mixin join  
+  response#join.
+    I would like to join House Turing!
+
+mixin deciding 
+  response#deciding.
+    I’m still deciding which house to join.
+
+mixin getting_started 
+  response#getting_started.
+    I’m looking forward to exploring Turing Fields too!
+
+conversation(display_name="Professor Taifa, Head of House Turing")
+  case lastResponse
+    
+    when "unique"
+      statement#unique
+        responses 
+          +join 
+          +deciding
+
+    when "join"
+      statement#join
+
+    when "deciding"
+      statement#deciding
+
+    when "getting_started"
+      statement#getting_started
+
+    default
+      if !worldState.playerHouse
+        - worldState.playerHouse = null;
+        statement#default
+        responses
+          +unique 
+          +join 
+          +deciding 
+
+      else
+        statement#gauntlet_default
+        responses
+            +getting_started
+
+  statements
+    statement#default
+      text.
+        A House Ceremony and the House Gauntlet taking place in the same week! 
+        It is certainly an exciting time to be at Arcane Academy.
+
+  statement#unique
+      text.
+        House Turing prides itself on its commitment to problem-solving and innovation. 
+        Students in House Turing enjoy participating in activities across the Academy, 
+        from codeball to extracurriculars to career events.
+
+  statement#join
+      text.
+        House Turing is lucky to have you! The last step in the House Ceremony 
+        involves confirming your choice by taking your pledge scroll to the 
+        #[span.highlight House Fire] at the center of the Hall.
+
+  statement#deciding
+      text.
+        By all means! As soon as you choose, you’ll be able to attend the House Gauntlet, 
+        which is one of Arcane’s most exciting annual events.
+
+  statement#gauntlet_default 
+    text.
+      House Turing was the first house to start the Gauntlet 
+      last year and the year before. Who knows, maybe once you’ve completed 
+      House Lovelace, House Turing might be selected as the second! 
+
+  statement#getting_started 
+    text.
+      Good luck in Lovelace Tower for now.
+
+// Flush world state changes, if any
+- setState(STATE_KEY, worldState);
