+++ levels/house_ceremony/maps/catacombs.png
