+++ levels/house_ceremony/objectives/hopper_key/objective.json
