+++ levels/house_ceremony/objectives/hopper_key/validator.js
