+++ levels/house_ceremony/objectives/hopper_key/walkthrough.md
