+++ levels/house_ceremony/objectives/lovelace_key/objective.json
