+++ levels/house_ceremony/objectives/lovelace_key/validator.js
