+++ levels/house_ceremony/objectives/neumann_key/objective.json
