+++ levels/house_ceremony/objectives/neumann_key/validator.js
