+++ levels/house_ceremony/objectives/neumann_key/walkthrough.md
