+++ levels/house_ceremony/objectives/obtain_wand/description.md
