+++ levels/house_ceremony/objectives/obtain_wand/objective.json
