+++ levels/house_ceremony/objectives/obtain_wand/validator.js
