+++ levels/house_ceremony/objectives/obtain_wand/walkthrough.md
