+++ levels/house_ceremony/objectives/turing_key/validator.js
