--- levels/house_ceremony_aa3/conversations/head-clay.pug
@@ -1,31 +1,0 @@
--
-  const STATE_KEY = 'TQ_API_ACADEMY_WORLD_STATE';
-  const worldState = getState(STATE_KEY);    
-
-mixin excavation
-  response#excavation.
-    Why does the Academy have an excavation crew?
-
-conversation(display_name="Professor Clay, Head of House Lovelace")
-  case lastResponse
-    
-    when "excavation"
-      statement#excavation
-
-    default
-      statement#default
-      responses
-        +excavation
-
-  statements
-    statement#default
-      text.
-        Oh, this is embarrassing! I'm sorry, but the way forward is blocked right now. 
-        I've sent a raven requesting the Academy excavation crew arrive as soon as possible.
-    statement#excavation
-      text.
-        Our Catacombs couldn't dig themselves! We have not needed the crew in some time, though. 
-        It will take a #[span.highlight new version of TwilioQuest] for them to get here. Please check back later!
-
-// Flush world state changes, if any
-- setState(STATE_KEY, worldState);
