+++ scripts/config.js
@@ -1,3 +1,5 @@
 module.exports = {
-  WORLD_STATE_KEY: "TQ_API_ACADEMY_WORLD_STATE",
+  PRE_ACADEMY_STATE_KEY: "TQ_API_ACADEMY_WORLD_STATE",
+  HOUSE_CEREMONY_STATE_KEY: "TQ_HOUSE_CEREMONY_WORLD_STATE",
+  LOVELACE_TOWER_STATE_KEY: "TQ_LOVELACE_TOWER_WORLD_STATE",
 };
